refactor: refactored modules and hir deffinitions
