+++ frontend/src/hir/model/mod.rs
@@ -1,0 +1,6 @@
+mod declarations;
+mod expression;
+mod statments;
+pub use declarations::*;
+pub use expression::*;
+pub use statments::*;
