--- frontend/src/hir/types/tys.rs
@@ -1,1 +1,0 @@
-
